Add option to control DTX (Discontinuous Transmission) (#49)
* Add option to control DTX (Discontinuous Transmission)

Add an option to `AddTrack` message to enable/disable DTX.

At the API level, the parameter is named `audioOptimizeBandwidth`
to hide the technical description of DTX from developers.
The setting defaults to true. If users do not want bandwidth
optimization (for e.g. for music kind of use cases), turning
off audioOptimizeBandwidth will disable DTX.

Testing:
--------
Tested on both Chrome and Firefox and can see up to 70% saving
when not talking even if the mic is unmuted and there is some
background noise.

* Revert test code

* Remove debug code

* audioOptimizeBandwidth -> dtx per David's suggestion.