Fix simulcast layer (#60)
* Fix simulcast layer

Problem:
--------
Simulcast encoding set up was using [f, h, q] or [f, h]
depending on request resolution. That resulted in two issues
- ion-sfu has hard mapping to those rids. So, in the second
case, there is no low quality layer. Actually, it probably
does not matter as it is relative, but we have stream tracker
not running on the lowest layer. So, in the second case,
the stream tracker is running on both layers although there
are only two layers.
- More importantly, Chrome seems to be stupid (have not checked
Firefox). It seems to be using index into the encoding array
to decide which layer to stop when there are constraints.
So, in the [f, h, q] case, it is actually stopping the `q` layer
as it is last in the list and Chrome thinks that is the highest
quality (maybe that ordering is a standard and that is how it is
supposed to work, but I can't find documentation)

Fix:
----
- Use ordering of [q, h, f] or [q, h] to satisfy both sides
- Removed some code which were ordering stats based on rid order.
As stats map is used with rid key, the ordering is not important.
- Also changed the bestEncoding determination to not rely on ordering
of encodings.

Testing:
--------
- Ensure that when Chrome stops a layer due to bandwidth (oy maybe
CPU limitations, must be CPU in my case as I have a slow machine),
it stops the highest spatial layer and not the lowest.

* Add comments

* Address PR comments about a comment :-)