Default case for createLocalAudio/VideoTrack
Hey LiveKit team! First of all, upgrading from v0.14 to v0.15 was super easy, so nice work on that release.

I noticed the previous `createLocalVideoTrack` behavior didn't require any arguments to actually get a video track with the default camera. It seemed weird to have to add `createLocalVideoTrack(true)`, so I just made this PR to fallback to that for these helper functions.