Handle device mute and document freeze events (#734)
When a source is unable to produce, it'll emit `mute` events. We are
currently not capturing these and it could lead to unstable/garbage data
being published to the SFU.

This PR will generate a publisher mute event when it is unable to produce.
It also consolidates the multiple places where we store the
_mediaStreamTrack reference to ensure consistent handling.

We also handle `freeze` event emitted on the document. This takes place
in Chrome when memory saver is activated. Google recommends tearing down
WebRTC connections: https://developer.chrome.com/blog/page-lifecycle-api/#developer-recommendations-for-each-state