Treat all signal messages as ping
In cases of rooms that are highly active, it's possible to have a lot of
signal messages queued that it blocks the ping messages for awhile.

Since the purpose of ping/pong is to ensure signal channel is active, it
makes sense to use any type of signal traffic as that indicator of liveliness.