Fix FPS and latency bugs with VP9 screenshare.
Also restored L3T3_KEY publishing for screen share. It was disabled previously.

Chrome does not allow more than 5 fps with L1T3, and it has encoding bugs with L3T3
It has a different path for screenshare handling and it seems to be untested/buggy
As a workaround, we are setting contentHint to force it to go through the same
path as regular camera video. While this is not optimal, it delivers the performance
that we need