Fix quality issues with SVC screenshares. (#1077)
When we munged SDP to set bitrate, it was set on the codec's fmtp line.
The codec and its fmtp is shared across media tracks. So if a user is sharing
a camera and screenshare tracks, both tracks will be published at the
same bitrate (typically the user's camera track). This leads to blurry
screenshares.

Previously this was set because AV1 took a long time to start up. So we are
keeping that behavior only for AV1. I've confirmed that VP9 does not suffer
from the slow-start.

Additionally, screenshares cannot be encoded in SVC with L3T3_KEY. Chrome
degrades quality aggressively whenever L1T3 isn't used.