fix: add missing async
Previously, there was an error when `await postAction()` was being run
later on because when postAction was being defined, it wasn't made
explicit that it always returned a promise, so typescript thought that
`await undefined` was effectively running sometimes. So, update the
function definition to make it clear it is always async.